requirements.txt
# MCP Server Dependencies
mcp>=1.0.0
httpx>=0.25.0
asyncio-mqtt>=0.16.0

# Optional dependencies for enhanced functionality
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Configuration dependencies
PyYAML>=6.0

# Semantic search dependencies (optional)
sentence-transformers>=2.2.2
numpy>=1.24.0
scikit-learn>=1.3.0
